feature: 规范国际化变量 去除alias配置
