--- frontend/src/utils/index.js
@@ -1,5 +1,0 @@
-import * as format from './format/index.js'
-
-export default {
-    ...format
-}
